pyproject.toml
[project]
name = "agentic-observability"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastmcp>=2.9.0",
  "gradio>=5.34.2",
  "llama-cloud-services>=0.6.36",
  "llama-index-embeddings-openai>=0.3.1",
  "llama-index-indices-managed-llama-cloud>=0.6.11",
  "llama-index-llms-openai>=0.4.7",
  "llama-index-observability-otel>=0.1.1",
  "llama-index-tools-mcp>=0.2.5",
  "mypy>=1.16.1",
  "opentelemetry-exporter-otlp-proto-http>=1.34.1",
  "pandas>=2.3.0",
  "pre-commit>=4.2.0",
  "psycopg-binary>=3.2.9",
  "psycopg2-binary>=2.9.10",
  "pytest>=8.4.1",
  "python-dotenv>=1.1.0",
  "sqlalchemy>=2.0.41",
  "websockets>=15.0.1"
]

[tool.mypy]
disable_error_code = ["import-not-found", "import-untyped"]

[tool.pytest.ini_options]
pythonpath = ["src"]
